Make panels easier to work with and nicer to look at
